swagger added
Added swegger on routes: Auth, Recipes.
Partly - OwnRecipes.
Fixed temp image delete

+++ controllers/authController.js
@@ -9,7 +9,7 @@ const jimp = require("jimp");
 const { SECRET_KEY } = process.env;
 
 const { ctrlWrapper } = require("../helpers/index");
-const {HttpError} = require('../helpers')
+const { HttpError } = require("../helpers");
 const storeImage = path.resolve("public", "avatars");
 const { uploadImage, updateLoadedImage } = require("../helpers/cloudinary");
 
@@ -86,6 +86,7 @@ const userLogin = async (req, res, next) => {
 
 const userLogout = async (req, res, next) => {
   const { _id } = req.user;
+  console.log("req user", req.user);
   await User.findByIdAndUpdate(_id, { token: "" });
 
   res.json({
@@ -101,34 +102,36 @@ const userCurrent = async (req, res, next) => {
 
 const userUpdateAvatar = async (req, res, next) => {
   const { id } = req.user;
-  const { path: temporaryName, filename: newFileName } = req.file;
+  const user = await User.findById(id);
   const { name } = req.body;
+  const updName = name.trim() === "" ? user.name : name
+  let result = {};
+  if (req.file) {
+    const { path: temporaryName } = req.file;
+    try {
+      if (user.avatar) {
+        const { public_id } = user.avatar;
 
-  //   const fileName = path.join(storeImage, newFileName);
+        result = await updateLoadedImage(temporaryName, public_id);
+      } else {
+        result = await uploadImage(temporaryName);
+      }
 
-  try {
-    const user = await User.findById(id);
-    let result = null;
-    if (user.avatar) {
-      const { public_id } = user.avatar;
-
-      result = await updateLoadedImage(temporaryName, public_id);
-    } else {
-      result = await uploadImage(temporaryName);
+      fs.unlink(temporaryName);
+    } catch (error) {
+      fs.unlink(temporaryName);
+      next(error);
     }
-
-    await User.findByIdAndUpdate(id, {
-      name,
-      avatarURL: result.url,
-      avatar: result,
-    });
-    fs.unlink(temporaryName);
-    res.status(200);
-    res.json(`avatar changed to ${result.url}, name changed to ${name}`);
-  } catch (error) {
-    fs.unlink(temporaryName);
-    next(error);
   }
+
+  await User.findByIdAndUpdate(id, {
+    name: updName,
+    avatarURL: result.hasOwnProperty('url') ? result.url : user.avatarURL,
+    avatar: result.hasOwnProperty('url') ? result : user.avatar,
+  });
+
+  res.status(200);
+  res.json(`Avatar url is: ${result.hasOwnProperty('url') ? result.url : user.avatarURL}, username is ${updName}`);
 };
 
 module.exports = {
